tests renderselect without creation

+++ test/conftest.py
@@ -9,25 +9,33 @@ import pytest
 import os
 import time
 import subprocess
+import requests
 
 from dashUI import params
+from dashUI.start_webUI import prefix
 
 @pytest.fixture(scope='session')
 def startup_webui():
-    outfile = os.path.join(params.base_dir,'test','webui_temp.out')
+    response = requests.get(prefix + 'localhost:8050', verify=False)
 
-    p = subprocess.Popen(os.path.join(params.base_dir, 'WebUI.sh'), stdout=open(outfile, 'w'))
+    if response.status_code == 200:
+        yield 'server is already running'
+    else:
+        outfile = os.path.join(params.base_dir,'test','webui_temp.out')
 
-    # check for regular running
-    time.sleep(8)
+        p = subprocess.Popen(os.path.join(params.base_dir, 'WebUI.sh'), stdout=open(outfile, 'w'))
 
-    assert p.errors is None
+        # check for regular running
+        time.sleep(8)
 
-    assert p.returncode is None
-    assert os.path.exists(outfile)
+        assert p.errors is None
 
-    yield outfile
+        assert p.returncode is None
+        assert os.path.exists(outfile)
 
-    os.system('kill $(echo $(ps x -o "pid command" | grep "dashUI/app" |grep "sh -c python") | cut -d " " -f 1)')
-    os.system('kill $(echo $(ps x -o "pid command" | grep "dashUI/app" |grep "python") | cut -d " " -f 1)')
-    os.system('rm ' + outfile)
+        yield outfile
+
+    if not response.status_code == 200:
+        os.system('kill $(echo $(ps x -o "pid command" | grep "dashUI/app" |grep "sh -c python") | cut -d " " -f 1)')
+        os.system('kill $(echo $(ps x -o "pid command" | grep "dashUI/app" |grep "python") | cut -d " " -f 1)')
+        os.system('rm ' + outfile)
